Folder File encoder was renamed to Simple File Encryptor
And in the files, a lot of changes are made.
Sorry about writing it like this, but I can't commit the renaming without the changes.
Basically, the only thing that needs to be done is implementing methods in the Encryptor class
And also lots of other things I forgot about...

Sorry again for this commit...