Fixed an error with OpenFileDialog after selecting and invalid file
When a user selects an .ecp file for encrypting or non-.ecp file for decryptning, a message box asking if the user wants to perform the opposite action displays. If the user clicks NO, the dialog isn't closed and the user can select another file.
Before, a new dialog opened overe the first one and after selecting a valid file, an exception was thrown.